pyproject.toml
[project]
name = "BootCamp_Codigo_Facilito"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Fernando Bardelli", email = "fernando@fliabardelli.com.ar"},
]
dependencies = ["django>=5.2.1", "djangorestframework>=3.16.0", "django-cors-headers>=4.7.0", "djangorestframework-simplejwt>=5.5.0", "django-extensions>=4.1", "gunicorn>=23.0.0", "drf-spectacular[sidecar]>=0.28.0", "whitenoise>=6.9.0", "django-environ>=0.12.0", "django-debug-toolbar>=5.2.0"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
